Changes to the way that the Write File operates.  You now can send it to the selected collection in the File View portion of the display.
This added the PutEvent and the PutListener.  These just distinguish
when a Put occurs to either the selected collection on the Tree View,
or on the file View.

The way selection occurs has changed.  There is now a select string in Extra.
There was a pervasive bug before which set variables in
the wrong thread.  Also a "processed exception" has been changed so as to
alert the user of the exception.

The current state is buggy since rename sometimes does not work.  This
is because the getSelectedResouce returns an invalid (usually an old
string) value.  This may cause other bugs to appear as well.

A pervasive bug in handle response may be related to these resource problems.
What happens is that the getStatus call hangs in try.