Added checks to the response to MOVE and DELETE.  The reason for this is to guard for cases involving a locked item.  The lock token sent may not be valid, or the item could have reset the token (unlock). This resonds as a 412 or 423 on either MOVE or DELETE.
The ResponseInterpreter then does a DiscoverLock and on that return,
attemps the action for a second time.  If this fails, the user is notified,
(cases for trying to delete some sles's locked item) and in the case of
a Rename, the old value is reset in the Table.

The RequestGenerator has some new items which save the state of the first
attempt.  This is needed to ensure that the second attempt will have the
proper values in it.